enforce auth for franchises

+++ src/routes/franchiseRouter.js
@@ -93,12 +93,17 @@ franchiseRouter.post(
   })
 );
 
-//TODO check auth. Anyone can delete a franchise.
 // deleteFranchise
 franchiseRouter.delete(
   '/:franchiseId',
+  authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
+    const franchise = await DB.getFranchise({ id: franchiseId });
+    if (!franchise || (!req.user.isRole(Role.Admin) && !franchise.admins.some((admin) => admin.id === req.user.id))) {
+      throw new StatusCodeError('unable to delete a franchise', 403);
+    }
+
     await DB.deleteFranchise(franchiseId);
     res.json({ message: 'franchise deleted' });
   })
